+++ src/fobi/contrib/plugins/form_elements/content/content_image/fobi_form_elements.py
@@ -81,10 +81,7 @@ class ContentImagePlugin(FormElementPlugin):
             'label': '',
         }
 
-        form_field_instances = []
-
-        form_field_instances.append((self.data.name, NoneField, kwargs))
-        return form_field_instances
+        return [(self.data.name, NoneField, kwargs)]
 
 
 form_element_plugin_registry.register(ContentImagePlugin)