+++ src/fobi/contrib/plugins/form_elements/content/content_video/fobi_form_elements.py
@@ -46,10 +46,7 @@ class ContentVideoPlugin(FormElementPlugin):
             'label': '',
         }
 
-        form_field_instances = []
-
-        form_field_instances.append((self.data.name, NoneField, kwargs))
-        return form_field_instances
+        return [(self.data.name, NoneField, kwargs)]
 
 
 form_element_plugin_registry.register(ContentVideoPlugin)