+++ src/fobi/contrib/plugins/form_elements/fields/select_multiple_with_max/forms.py
@@ -3,6 +3,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from fobi.base import BaseFormFieldPluginForm, get_theme
 from fobi.helpers import validate_initial_for_multiple_choices
+from fobi.widgets import NumberInput
 
 __title__ = 'fobi.contrib.plugins.form_elements.fields.select_multiple.forms'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
@@ -95,7 +96,7 @@ class SelectMultipleWithMaxInputForm(forms.Form, BaseFormFieldPluginForm):
     max_choices = forms.IntegerField(
         label=_("Max choices"),
         required=False,
-        widget=forms.widgets.NumberInput(
+        widget=NumberInput(
             attrs={'class': theme.form_element_html_class}
         )
     )