+++ src/fobi/contrib/plugins/form_elements/fields/slider/defaults.py
@@ -1,0 +1,16 @@
+__title__ = 'fobi.contrib.plugins.form_elements.fields.' \
+            'slider.defaults'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2014-2016 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = (
+    'INITIAL',
+    'MAX_VALUE',
+    'MIN_VALUE',
+    'STEP',
+)
+
+INITIAL = 50
+MIN_VALUE = 0
+MAX_VALUE = 100
+STEP = 1