+++ src/fobi/contrib/themes/bootstrap3/widgets/form_elements/slider_bootstrap3_widget/static/bootstrap3/js/fobi.plugin.slider-bootstrap3-widget.js
@@ -1,0 +1,12 @@
+/*
+    Document   : fobi.plugin.slider-bootstrap3-widget.js
+    Created on : Oct 19, 2016, 03:30:35 AM
+    Author     : Artur Barseghyan (artur.barseghyan@gmail.com)
+    Description:
+        `bootstrap3` theme slider plugin scripts.
+*/
+;
+
+$(document).ready(function() {
+    $('.slider').bootstrapSlider();
+});