+++ src/fobi/migrations/0010_formwizardhandler.py
@@ -5,14 +5,21 @@ from __future__ import unicode_literals
 from django.conf import settings
 from django.db import migrations, models
 
+from nine import versions
 
 class Migration(migrations.Migration):
 
-    dependencies = [
-        ('auth', '0007_alter_validators_add_error_messages'),
-        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
-        ('fobi', '0009_formwizardentry_wizard_type'),
-    ]
+    if versions.DJANGO_LTE_1_9:
+        dependencies = [
+            migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+            ('fobi', '0009_formwizardentry_wizard_type'),
+        ]
+    else:
+        dependencies = [
+            ('auth', '0007_alter_validators_add_error_messages'),
+            migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+            ('fobi', '0009_formwizardentry_wizard_type'),
+        ]
 
     operations = [
         migrations.CreateModel(